Refactor Expander to remove components level implementation.
